Improve detection of RPN expressions with a better regex.
If you had a line of text like this, with | being the cursor position:

My answer is 3 14 *|

this plugin was unable to detect the RPN expression. It would find the
`e`, but because the text after it (`r is `) was invalid, the expression
wouldn't be processed. To fix this, I required the regex to match up to
the cursor position. Now with the same line of text, it successfully
matches `3 14 *`.

Because it matches up to the cursor position, and not the end of the
line, the plugin will also process these lines:

3 14 *| is my answer.
For the last time 3 14 *| is my answer.

The only caveat is that the cursor must be at the end of the RPN
expression.

I also had to make sure it was looking at complete operator words;
otherwise the last line's expression being evaluated would have been
`e 3 14 *`