Create the user command for evaluating an RPN expression.
For `:RPN` and `:RPN!`
The current line or visual selection is replaced by result of the
line(s)'s RPN expression. If the bang symbol is used, the expression
itself is also placed in the output.

For `:RPN <rpn expression>` and `:RPN! <rpn expression>`
The result of the RPN expression is inserted into the line above the
current line. If the bang symbol is used, the expression itself is also
placed in the output.