Standalone command (#12)
* Split the RPN calculation code out from the nvim_cmp source code.

* Move tests to subdirectory, and reconfigure the tests.

Most tests were rewritten to go against the rpn.lua code. These were the
ones that test the operators' functions. Other tests, ones that don't
have operators, ones that return errors (like division by 0), and ones
that don't affect the stack, remain as a test against the code that
interfaces with nvim_cmp.

* Create the user command for evaluating an RPN expression.

For `:RPN` and `:RPN!`
The current line or visual selection is replaced by result of the
line(s)'s RPN expression. If the bang symbol is used, the expression
itself is also placed in the output.

For `:RPN <rpn expression>` and `:RPN! <rpn expression>`
The result of the RPN expression is inserted into the line above the
current line. If the bang symbol is used, the expression itself is also
placed in the output.

* Change the test for `i` to be more in line with the new assertCMP.

* Break a line of code for better clarity. Replace vim.fn.join().

* Update documentation with the new `:RPN` command.